Extract AnalysisScaffodling from TestHelper
